feat: replication中fdtp相关代码根据配置判断是否启动 #1037
