+++ src/backend/replication/biz-replication/src/main/kotlin/com/tencent/bkrepo/replication/replica/replicator/base/internal/ClusterArtifactReplicationHandler.kt
@@ -27,13 +27,10 @@
 
 package com.tencent.bkrepo.replication.replica.replicator.base.internal
 
-import com.tencent.bkrepo.common.api.constant.HttpStatus
 import com.tencent.bkrepo.common.api.constant.StringPool
 import com.tencent.bkrepo.common.artifact.stream.rateLimit
 import com.tencent.bkrepo.common.storage.pojo.FileInfo
 import com.tencent.bkrepo.common.artifact.util.http.StreamRequestBody
-import com.tencent.bkrepo.fdtp.codec.DefaultFdtpHeaders
-import com.tencent.bkrepo.fdtp.codec.FdtpResponseStatus
 import com.tencent.bkrepo.replication.config.ReplicationProperties
 import com.tencent.bkrepo.replication.constant.BLOB_PUSH_URI
 import com.tencent.bkrepo.replication.constant.BOLBS_UPLOAD_FIRST_STEP_URL_STRING
@@ -43,32 +40,24 @@ import com.tencent.bkrepo.replication.constant.SHA256
 import com.tencent.bkrepo.replication.constant.SIZE
 import com.tencent.bkrepo.replication.constant.STORAGE_KEY
 import com.tencent.bkrepo.replication.enums.WayOfPushArtifact
-import com.tencent.bkrepo.replication.exception.ArtifactPushException
-import com.tencent.bkrepo.replication.fdtp.FdtpAFTClientFactory
-import com.tencent.bkrepo.replication.fdtp.FdtpServerProperties
 import com.tencent.bkrepo.replication.manager.LocalDataManager
 import com.tencent.bkrepo.replication.pojo.blob.RequestTag
-import com.tencent.bkrepo.replication.replica.base.process.ProgressListener
 import com.tencent.bkrepo.replication.replica.context.FilePushContext
 import com.tencent.bkrepo.replication.replica.replicator.base.ArtifactReplicationHandler
-import io.netty.channel.ChannelProgressiveFuture
-import io.netty.channel.ChannelProgressiveFutureListener
 import okhttp3.MultipartBody
 import okhttp3.Request
 import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Lazy
 import org.springframework.cloud.context.config.annotation.RefreshScope
 import org.springframework.stereotype.Component
-import java.net.ConnectException
-import java.util.concurrent.TimeUnit
 
 @RefreshScope
 @Component
 class ClusterArtifactReplicationHandler(
     localDataManager: LocalDataManager,
     replicationProperties: ReplicationProperties,
-    val fdtpServerProperties: FdtpServerProperties,
-    val listener: ProgressListener
 ) : ArtifactReplicationHandler(localDataManager, replicationProperties) {
 
     //不支持yaml list配置 https://github.com/spring-projects/spring-framework/issues/16381
@@ -80,6 +69,10 @@ class ClusterArtifactReplicationHandler(
 
     @Value("\${replication.httpRepos:}")
     private var httpRepos: List<String> = emptyList()
+
+    @Autowired
+    @Lazy
+    private lateinit var fdtpPusher: FdtpPusher
 
     override fun blobPush(
         filePushContext: FilePushContext,
@@ -98,7 +91,8 @@ class ClusterArtifactReplicationHandler(
                 super.blobPush(filePushContext, newType, downGrade)
             }
             WayOfPushArtifact.PUSH_WITH_FDTP.value -> {
-                pushWithFdtp(filePushContext)
+                require(::fdtpPusher.isInitialized) { "fdtp is disabled" }
+                fdtpPusher.pushBlob(filePushContext)
             }
             else -> {
                 pushBlob(filePushContext)
@@ -203,69 +197,6 @@ class ClusterArtifactReplicationHandler(
         }
     }
 
-    /**
-     * 使用fdtp推送blob文件数据到远程集群
-     */
-    private fun pushWithFdtp(filePushContext: FilePushContext): Boolean {
-        with(filePushContext) {
-            logger.info("File $sha256 will be pushed using the fdtp way.")
-            val client = FdtpAFTClientFactory.createAFTClient(context.cluster, fdtpServerProperties.port)
-            val artifactInputStream = localDataManager.getBlobData(sha256!!, size!!, context.localRepo)
-            val rateLimitInputStream = artifactInputStream.rateLimit(
-                replicationProperties.rateLimit.toBytes()
-            )
-            val storageKey = context.remoteRepo?.storageCredentials?.key
-            val headers = DefaultFdtpHeaders()
-            headers.add(SHA256, sha256)
-            storageKey?.let { headers.add(STORAGE_KEY, storageKey) }
-            try {
-                val progressListener = object : ChannelProgressiveFutureListener {
-                    private val tag = RequestTag(context.task, sha256, size)
-                    private val progressListener: ProgressListener = listener
-                    private var previous: Long = 0
-
-                    @Throws(Exception::class)
-                    override fun operationProgressed(future: ChannelProgressiveFuture?, progress: Long, total: Long) {
-                        if (progress == previous) return
-                        try {
-                            progressListener.onProgress(tag.task, tag.key, progress - previous)
-                        } catch (ignore: Exception) {
-                        }
-                        previous = progress
-                    }
-
-                    @Throws(Exception::class)
-                    override fun operationComplete(future: ChannelProgressiveFuture) {
-                        if (future.isSuccess) {
-                            progressListener.onSuccess(tag.task)
-                        } else {
-                            progressListener.onFailed(tag.task, tag.key)
-                        }
-                    }
-                }
-                listener.onStart(context.task, sha256,0)
-
-                val responsePromise = client.sendStream(rateLimitInputStream, headers, progressListener)
-
-                val response = responsePromise.get(READ_TIME_OUT, TimeUnit.SECONDS)
-                if (response.status == FdtpResponseStatus.OK){
-                    return true
-                } else {
-                    val logMessage = "Error occurred while pushing file $sha256 " +
-                        "with the fdtp way, error is ${response.status.reasonPhrase}"
-                    logger.warn(logMessage)
-                    throw ArtifactPushException(logMessage)
-                }
-            } catch (e: ConnectException) {
-                // 当不支持fdtp方式进行传输时抛出异常，进行降级处理
-                logger.warn(
-                    "Error occurred while pushing file $sha256 with the fdtp way, errors is ${e.message}", e
-                )
-                throw ArtifactPushException(e.message.orEmpty(), HttpStatus.METHOD_NOT_ALLOWED.value)
-            }
-        }
-    }
-
     private fun buildParams(
         sha256: String,
         filePushContext: FilePushContext,
@@ -318,7 +249,5 @@ class ClusterArtifactReplicationHandler(
 
     companion object {
         private val logger = LoggerFactory.getLogger(ClusterArtifactReplicationHandler::class.java)
-        // 读取结果返回超时时间 15分钟
-        private const val READ_TIME_OUT = 60L * 15
     }
 }
